데이터 받아오는 로직 수정, category 추가,  network, 거리 계산 로직 추가

+++ app/src/main/java/kr/ac/hansung/gyunggilocalmoneymap/data/remote/source/OpenApiRemoteDataSource.kt
@@ -6,7 +6,7 @@ import kr.ac.hansung.gyunggilocalmoneymap.data.remote.model.SHPlace
 
 interface OpenApiRemoteDataSource {
 
-    val pageLoadingSubject: BehaviorSubject<Float>
+    val pageLoadingSubject: BehaviorSubject<Int>
 
     fun getPlacesByIndex(pIndex: String) : Single<List<SHPlace>>
